FIX missing Driver References in Dot MATRIX project
Missing references to timer and adc, we added them to the project file
and added them to intialization functions.
Note: The partial fix we did to ADC definitons, we just added the ports
for the ADC because we were unsure what they were suppose to be so we
just put random ones. Please fix.